+++ app/src/main/java/app/passwordstore/ui/sshkeygen/SshKeyGenActivity.kt
@@ -117,7 +117,7 @@ class SshKeyGenActivity : AppCompatActivity() {
               suspendCoroutine { cont ->
                 BiometricAuthenticator.authenticate(
                   this@SshKeyGenActivity,
-                  R.string.biometric_prompt_title_ssh_keygen,
+                  dialogSubTitleRes = R.string.biometric_prompt_title_ssh_keygen,
                   allowPin = true,
                 ) { result ->
                   // Do not cancel on failed attempts as these are handled by the